+++ crates/tytanic/src/cli/util/migrate.rs
@@ -1,12 +1,14 @@
-use std::collections::{BTreeMap, BTreeSet};
+use std::collections::BTreeMap;
 use std::fs;
 use std::io::Write;
-use std::path::Path;
+use std::path::PathBuf;
 
-use color_eyre::eyre::{self, WrapErr};
+use color_eyre::eyre;
 use termcolor::Color;
 use tytanic_core::project::Paths;
-use tytanic_core::test::Id;
+use tytanic_core::stdx;
+use tytanic_core::test::{Id, Suite};
+use tytanic_core::test_set::{eval, TestSet};
 
 use crate::cli::Context;
 use crate::{cwrite, ui};
@@ -25,12 +27,15 @@ pub struct Args {
 
 pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
     let project = ctx.project()?;
+    let suite = Suite::collect(
+        project.paths(),
+        &TestSet::new(eval::Context::empty(), eval::Set::built_in_all()),
+    )?;
+
     let paths = project.paths();
     let mut w = ctx.ui.stderr();
 
-    let mappings = collect_old_structure(paths, &args.name)?;
-
-    if mappings.is_empty() {
+    if suite.nested().is_empty() {
         writeln!(w, "No tests need to be moved")?;
         return Ok(());
     }
@@ -39,6 +44,16 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
         writeln!(w, "Moving tests:")?;
     } else {
         writeln!(w, "These tests would be moved:")?;
+    }
+
+    let mut has_colission = false;
+    let mut mappings = BTreeMap::new();
+    for old in suite.nested().keys() {
+        let new = Id::new(format!("{old}/{}", args.name))?;
+        let collision = suite.matched().contains_key(&new) || suite.filtered().contains_key(&new);
+
+        has_colission |= collision;
+        mappings.insert(old.clone(), (new, collision));
     }
 
     for (old, (new, collision)) in &mappings {
@@ -56,17 +71,6 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
 
     writeln!(w)?;
 
-    let mut has_colission = false;
-    for (old, (new, collision)) in &mappings {
-        if !*collision {
-            let old = paths.test_dir(old);
-            let new = paths.test_dir(new);
-            std::fs::rename(old, new)?;
-        } else {
-            has_colission = true;
-        }
-    }
-
     if has_colission {
         let mut w = ctx.ui.hint()?;
         cwrite!(bold_colored(w, Color::Red), "*")?;
@@ -79,7 +83,15 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
         writeln!(w)?;
     }
 
-    if !args.confirm {
+    if args.confirm {
+        for (old, (new, collision)) in &mappings {
+            if *collision {
+                continue;
+            }
+
+            migrate_test(paths, old, new)?;
+        }
+    } else {
         writeln!(ctx.ui.warn()?, "Make sure to back up your code!")?;
 
         {
@@ -107,60 +119,34 @@ pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
     Ok(())
 }
 
-pub fn collect_old_structure(
+// NOTE(tinger): I have no idea why simply renaming the test directory doesn't
+// work, but renaming the ref directory works
+
+fn migrate_test_part(
     paths: &Paths,
-    migration_name: &str,
-) -> eyre::Result<BTreeMap<Id, (Id, bool)>> {
-    let mut entries = BTreeSet::new();
-    for entry in paths.test_root().read_dir()? {
-        let entry = entry?;
-        if entry.metadata()?.is_dir() {
-            collect_old_structure_inner(paths, &entry.path(), &mut entries)?;
-        }
+    old: &Id,
+    new: &Id,
+    f: fn(&Paths, &Id) -> PathBuf,
+) -> eyre::Result<()> {
+    let old = f(paths, old);
+    let new = f(paths, new);
+
+    if old.try_exists()? {
+        fs::rename(&old, &new)?;
     }
 
-    let mut mappings = BTreeMap::new();
-    'outer: for id in &entries {
-        'inner: for internal in id.ancestors().skip(1) {
-            if !entries.contains(internal) {
-                continue 'inner;
-            }
-
-            let old = Id::new(internal)?;
-            let new = Id::new(format!("{internal}/{migration_name}"))?;
-            let colission = entries.contains(&new);
-
-            if mappings.insert(old, (new, colission)).is_some() {
-                continue 'outer;
-            }
-        }
-    }
-
-    Ok(mappings)
+    Ok(())
 }
 
-fn collect_old_structure_inner(
-    paths: &Paths,
-    path: &Path,
-    entries: &mut BTreeSet<Id>,
-) -> eyre::Result<()> {
-    if path.join("test.typ").try_exists()? {
-        entries.insert(Id::new_from_path(path.strip_prefix(paths.test_root())?)?);
-    }
-
-    for entry in fs::read_dir(path).wrap_err_with(|| format!("{path:?}"))? {
-        let entry = entry?;
-        let path = entry.path();
-        let name = entry.file_name();
-
-        if name == "ref" || name == "out" || name == "diff" {
-            continue;
-        }
-
-        if entry.metadata()?.is_dir() {
-            collect_old_structure_inner(paths, &path, entries)?;
-        }
-    }
-
+fn migrate_test(paths: &Paths, old: &Id, new: &Id) -> eyre::Result<()> {
+    let test_dir = paths.test_dir(new);
+    stdx::fs::create_dir(&test_dir, true)?;
+    migrate_test_part(paths, old, new, Paths::test_script)?;
+    migrate_test_part(paths, old, new, Paths::test_ref_script)?;
+    migrate_test_part(paths, old, new, Paths::test_ref_dir)?;
+    let out_dir = paths.test_out_dir(old);
+    stdx::fs::remove_dir(&out_dir, true)?;
+    let diff_dir = paths.test_diff_dir(old);
+    stdx::fs::remove_dir(&diff_dir, true)?;
     Ok(())
 }